connected login logic to frontend
